sys-apps/khph: Redigest (bgo#902313); update SRC_URI and copyright.
This ebuild was using deprecated digests in the Manifest (RMD160,
WHIRLPOOL), so the digests have been regenerated to stay current.

Also the explicit SRC_URI was preventing downloading from the regular
Hackage URL, requiring it to exist on the Gentoo mirror.  Removing
SRC_URI and using the eclass default fixes this.

Also, fix the copyright lines.